requirements.txt
pandas
numpy
numpy-financial
scipy
ipykernel
nbformat
matplotlib
seaborn
plotly
graphviz
PyYAML
# Jupyter ecosystem for Quarto
jupyter
jupyter-cache
nbclient
jupyter-client
jupyter-core
traitlets
notebook
nbconvert
ipywidgets
jupyterlab
pyairtable
python-dotenv
tabulate
PyYAML
psutil
PyPDF2
moviepy
openai-whisper
google-genai
anthropic
pydub
Pillow
